requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
attrs==25.3.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
choreographer==1.2.0
click==8.2.1
cmdstanpy==1.3.0
contourpy==1.3.3
cycler==0.12.1
decorator==5.2.1
distro==1.9.0
executing==2.2.1
filelock==3.19.1
fonttools==4.59.2
fsspec==2025.7.0
gitdb==4.0.12
GitPython==3.1.45
h11==0.16.0
hdbscan==0.8.40
hf-xet==1.1.9
holidays==0.83
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
idna==3.10
importlib_resources==6.5.2
iniconfig==2.3.0
ipython==9.6.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpickle==4.1.1
jsonschema==4.25.1
jsonschema-specifications==2025.4.1
kaleido==1.1.0
kiwisolver==1.4.9
logistro==2.0.1
MarkupSafe==3.0.2
matplotlib==3.10.6
matplotlib-inline==0.2.1
mpmath==1.3.0
narwhals==2.2.0
networkx==3.5
nltk==3.9.1
numpy==2.3.2
openai==2.6.1
orjson==3.11.4
packaging==25.0
pandas==2.3.2
parso==0.8.5
pexpect==4.9.0
pillow==11.3.0
plotly==6.3.0
pluggy==1.6.0
prompt_toolkit==3.0.52
prophet==1.2.1
protobuf==6.32.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pydantic==2.12.3
pydantic_core==2.41.4
pydeck==0.9.1
pygame==2.6.1
Pygments==2.19.2
pyparsing==3.2.3
pytest==8.4.2
pytest-timeout==2.4.0
python-dateutil==2.9.0.post0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.8.29
reportlab==4.4.3
requests==2.32.5
rpds-py==0.27.1
safetensors==0.6.2
scikit-learn==1.7.1
scipy==1.16.1
seaborn==0.13.2
sentence-transformers==5.1.2
setuptools==80.9.0
simplejson==3.20.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
stanio==0.5.1
streamlit==1.49.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.0
toml==0.10.2
torch==2.8.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.56.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
wcwidth==0.2.14
wordcloud==1.9.4
